chore: move oauth client code into folder
