+++ examples/oauth/decode-invoice.js
@@ -1,6 +1,7 @@
 import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
-import { auth, Client } from "../../dist/index.module.js";
+import { oauth } from "../../dist/index.module.js";
+const { auth, Client } = oauth;
 
 const rl = readline.createInterface({ input, output });
