+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@getalby/sdk",
-  "version": "4.1.1",
+  "version": "5.0.0",
   "description": "The SDK to integrate with Nostr Wallet Connect and the Alby API",
   "repository": "https://github.com/getAlby/js-sdk.git",
   "bugs": "https://github.com/getAlby/js-sdk/issues",
