+++ src/nwc/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
 export * from "./NWCClient";
+export * from "./NWAClient";
 export * from "./NWCWalletService";
 export * from "./NWCWalletServiceRequestHandler";
