+++ src/oauth/AlbyResponseError.ts
@@ -1,12 +1,15 @@
 export class AlbyResponseError extends Error {
   status: number;
   statusText: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   headers: Record<string, any>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   error: any; // todo: typeable?
   constructor(
     status: number,
     statusText: string,
     headers: Headers,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     error: any,
   ) {
     let message = status.toString();
