+++ src/webln/OauthWeblnProvider.ts
@@ -1,5 +1,5 @@
-import { Client } from "../client";
-import { OAuthClient, KeysendRequestParams } from "../types";
+import { Client } from "../oauth/client";
+import { OAuthClient, KeysendRequestParams } from "../oauth/types";
 
 interface RequestInvoiceArgs {
   amount: string | number;
